Modified add_order and delete_order functions
I have made modifications to the 'add_order' function to ensure that the quantity of a given product available for sale is decremented by the quantity of that product added to the orders table. I have also updated the 'delete_order' function to incorporate a feature that enables the addition of the quantity of a returned product to the existing quantity of that same product in the 'Products' table.